webrtc::error

Enum Error

#[non_exhaustive]
pub enum Error {

    ErrUnknownType,
    ErrConnectionClosed,
    ErrDataChannelNotOpen,
    ErrCertificateExpired,
    ErrNoTurnCredentials,
    ErrTurnCredentials,
    ErrExistingTrack,
    ErrPrivateKeyType,
    ErrModifyingPeerIdentity,
    ErrModifyingCertificates,
    ErrNonCertificate,
    ErrModifyingBundlePolicy,
    ErrModifyingRTCPMuxPolicy,
    ErrModifyingICECandidatePoolSize,
    ErrStringSizeLimit,
    ErrMaxDataChannelID,
    ErrNegotiatedWithoutID,
    ErrRetransmitsOrPacketLifeTime,
    ErrCodecNotFound,
    ErrNoRemoteDescription,
    ErrIncorrectSDPSemantics,
    ErrIncorrectSignalingState,
    ErrProtocolTooLarge,
    ErrSenderNotCreatedByConnection,
    ErrSenderInitialTrackIdAlreadySet,
    ErrSessionDescriptionNoFingerprint,
    ErrSessionDescriptionInvalidFingerprint,
    ErrSessionDescriptionConflictingFingerprints,
    ErrSessionDescriptionMissingIceUfrag,
    ErrSessionDescriptionMissingIcePwd,
    ErrSessionDescriptionConflictingIceUfrag,
    ErrSessionDescriptionConflictingIcePwd,
    ErrNoSRTPProtectionProfile,
    ErrFailedToGenerateCertificateFingerprint,
    ErrNoCodecsAvailable,
    ErrUnsupportedCodec,
    ErrSenderWithNoCodecs,
    ErrRTPSenderNewTrackHasIncorrectKind,
    ErrRTPSenderNewTrackHasIncorrectEnvelope,
    ErrRTPSenderDataSent,
    ErrRTPSenderSeqTransEnabled,
    ErrUnbindFailed,
    ErrNoPayloaderForCodec,
    ErrRegisterHeaderExtensionInvalidDirection,
    ErrRegisterHeaderExtensionNoFreeID,
    ErrSimulcastProbeOverflow,
    ErrDetachNotEnabled,
    ErrDetachBeforeOpened,
    ErrDtlsTransportNotStarted,
    ErrDtlsKeyExtractionFailed,
    ErrFailedToStartSRTP,
    ErrFailedToStartSRTCP,
    ErrInvalidDTLSStart,
    ErrNoRemoteCertificate,
    ErrIdentityProviderNotImplemented,
    ErrNoMatchingCertificateFingerprint,
    ErrUnsupportedFingerprintAlgorithm,
    ErrICEConnectionNotStarted,
    ErrICECandidateTypeUnknown,
    ErrICEInvalidConvertCandidateType,
    ErrICEAgentNotExist,
    ErrICECandidatesConversionFailed,
    ErrICERoleUnknown,
    ErrICEProtocolUnknown,
    ErrICEGathererNotStarted,
    ErrNetworkTypeUnknown,
    ErrSDPDoesNotMatchOffer,
    ErrSDPDoesNotMatchAnswer,
    ErrPeerConnSDPTypeInvalidValue,
    ErrPeerConnStateChangeInvalid,
    ErrPeerConnStateChangeUnhandled,
    ErrPeerConnSDPTypeInvalidValueSetLocalDescription,
    ErrPeerConnRemoteDescriptionWithoutMidValue,
    ErrPeerConnRemoteDescriptionNil,
    ErrPeerConnSingleMediaSectionHasExplicitSSRC,
    ErrPeerConnRemoteSSRCAddTransceiver,
    ErrPeerConnSimulcastMidRTPExtensionRequired,
    ErrPeerConnSimulcastStreamIDRTPExtensionRequired,
    ErrPeerConnSimulcastIncomingSSRCFailed,
    ErrPeerConnStatsCollectionFailed,
    ErrPeerConnAddTransceiverFromKindOnlyAcceptsOne,
    ErrPeerConnAddTransceiverFromTrackOnlyAcceptsOne,
    ErrPeerConnAddTransceiverFromKindSupport,
    ErrPeerConnAddTransceiverFromTrackSupport,
    ErrPeerConnSetIdentityProviderNotImplemented,
    ErrPeerConnWriteRTCPOpenWriteStream,
    ErrPeerConnTransceiverMidNil,
    ErrRTPReceiverDTLSTransportNil,
    ErrRTPReceiverReceiveAlreadyCalled,
    ErrRTPReceiverWithSSRCTrackStreamNotFound,
    ErrRTPReceiverForSSRCTrackStreamNotFound,
    ErrRTPReceiverForRIDTrackStreamNotFound,
    ErrRTPReceiverStateChangeInvalid {
        from: State,
        to: State,
    },
    ErrRTPSenderTrackNil,
    ErrRTPSenderStopped,
    ErrRTPSenderTrackRemoved,
    ErrRTPSenderRidNil,
    ErrRTPSenderNoBaseEncoding,
    ErrRTPSenderBaseEncodingMismatch,
    ErrRTPSenderRIDCollision,
    ErrRTPSenderNoTrackForRID,
    ErrRTPSenderNil,
    ErrRTPReceiverNil,
    ErrRTPSenderDTLSTransportNil,
    ErrRTPSenderSendAlreadyCalled,
    ErrRTPTransceiverCannotChangeMid,
    ErrRTPTransceiverSetSendingInvalidState,
    ErrRTPTransceiverCodecUnsupported,
    ErrSCTPTransportDTLS,
    ErrSDPZeroTransceivers,
    ErrSDPMediaSectionMediaDataChanInvalid,
    ErrSDPMediaSectionMultipleTrackInvalid,
    ErrSettingEngineSetAnsweringDTLSRole,
    ErrSignalingStateCannotRollback,
    ErrSignalingStateProposedTransitionInvalid {
        from: RTCSignalingState,
        applying: RTCSdpType,
        is_local: bool,
    },
    ErrStatsICECandidateStateInvalid,
    ErrICETransportNotInNew,
    ErrCertificatePEMFormatError,
    ErrSCTPNotEstablished,
    ErrClosedPipe,
    ErrInterceptorNotBind,
    ErrExcessiveRetries,
    ErrRTPTooShort,
    Util(Error),
    Ice(Error),
    Srtp(Error),
    Dtls(Error),
    Data(Error),
    Sctp(Error),
    Sdp(Error),
    Interceptor(Error),
    Rtcp(Error),
    Rtp(Error),
    Utf8(FromUtf8Error),
    RcGen(Error),
    MpscSend(String),
    ParseInt(ParseIntError),
    ParseUrl(ParseError),
    InvalidPEM(String),
    new(String),

}

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

ErrUnknownType

ErrUnknownType indicates an error with Unknown info.

ErrConnectionClosed

ErrConnectionClosed indicates an operation executed after connection has already been closed.

ErrDataChannelNotOpen

ErrDataChannelNotOpen indicates an operation executed when the data channel is not (yet) open.

ErrCertificateExpired

ErrCertificateExpired indicates that an x509 certificate has expired.

ErrNoTurnCredentials

ErrNoTurnCredentials indicates that a TURN server URL was provided without required credentials.

ErrTurnCredentials

ErrTurnCredentials indicates that provided TURN credentials are partial or malformed.

ErrExistingTrack

ErrExistingTrack indicates that a track already exists.

ErrPrivateKeyType

ErrPrivateKeyType indicates that a particular private key encryption chosen to generate a certificate is not supported.

ErrModifyingPeerIdentity

ErrModifyingPeerIdentity indicates that an attempt to modify PeerIdentity was made after PeerConnection has been initialized.

ErrModifyingCertificates

ErrModifyingCertificates indicates that an attempt to modify Certificates was made after PeerConnection has been initialized.

ErrNonCertificate

ErrNonCertificate indicates that there is no certificate

ErrModifyingBundlePolicy

ErrModifyingBundlePolicy indicates that an attempt to modify BundlePolicy was made after PeerConnection has been initialized.

ErrModifyingRTCPMuxPolicy

ErrModifyingRTCPMuxPolicy indicates that an attempt to modify RTCPMuxPolicy was made after PeerConnection has been initialized.

ErrModifyingICECandidatePoolSize

ErrModifyingICECandidatePoolSize indicates that an attempt to modify ICECandidatePoolSize was made after PeerConnection has been initialized.

ErrStringSizeLimit

ErrStringSizeLimit indicates that the character size limit of string is exceeded. The limit is hardcoded to 65535 according to specifications.

ErrMaxDataChannelID

ErrMaxDataChannelID indicates that the maximum number ID that could be specified for a data channel has been exceeded.

ErrNegotiatedWithoutID

ErrNegotiatedWithoutID indicates that an attempt to create a data channel was made while setting the negotiated option to true without providing the negotiated channel ID.

ErrRetransmitsOrPacketLifeTime

ErrRetransmitsOrPacketLifeTime indicates that an attempt to create a data channel was made with both options max_packet_life_time and max_retransmits set together. Such configuration is not supported by the specification and is mutually exclusive.

ErrCodecNotFound

ErrCodecNotFound is returned when a codec search to the Media Engine fails

ErrNoRemoteDescription

ErrNoRemoteDescription indicates that an operation was rejected because the remote description is not set

ErrIncorrectSDPSemantics

ErrIncorrectSDPSemantics indicates that the PeerConnection was configured to generate SDP Answers with different SDP Semantics than the received Offer

ErrIncorrectSignalingState

ErrIncorrectSignalingState indicates that the signaling state of PeerConnection is not correct

ErrProtocolTooLarge

ErrProtocolTooLarge indicates that value given for a DataChannelInit protocol is longer then 65535 bytes

ErrSenderNotCreatedByConnection

ErrSenderNotCreatedByConnection indicates remove_track was called with a rtp_sender::RTCRtpSender not created by this PeerConnection

ErrSenderInitialTrackIdAlreadySet

ErrSenderInitialTrackIdAlreadySet indicates a second call to RTCRtpSender::set_initial_track_id which is not allowed. Purely internal error, should not happen in practice.

ErrSessionDescriptionNoFingerprint

ErrSessionDescriptionNoFingerprint indicates set_remote_description was called with a SessionDescription that has no fingerprint

ErrSessionDescriptionInvalidFingerprint

ErrSessionDescriptionInvalidFingerprint indicates set_remote_description was called with a SessionDescription that has an invalid fingerprint

ErrSessionDescriptionConflictingFingerprints

ErrSessionDescriptionConflictingFingerprints indicates set_remote_description was called with a SessionDescription that has an conflicting fingerprints

ErrSessionDescriptionMissingIceUfrag

ErrSessionDescriptionMissingIceUfrag indicates set_remote_description was called with a SessionDescription that is missing an ice-ufrag value

ErrSessionDescriptionMissingIcePwd

ErrSessionDescriptionMissingIcePwd indicates set_remote_description was called with a SessionDescription that is missing an ice-pwd value

ErrSessionDescriptionConflictingIceUfrag

ErrSessionDescriptionConflictingIceUfrag indicates set_remote_description was called with a SessionDescription that contains multiple conflicting ice-ufrag values

ErrSessionDescriptionConflictingIcePwd

ErrSessionDescriptionConflictingIcePwd indicates set_remote_description was called with a SessionDescription that contains multiple conflicting ice-pwd values

ErrNoSRTPProtectionProfile

ErrNoSRTPProtectionProfile indicates that the DTLS handshake completed and no SRTP Protection Profile was chosen

ErrFailedToGenerateCertificateFingerprint

ErrFailedToGenerateCertificateFingerprint indicates that we failed to generate the fingerprint used for comparing certificates

ErrNoCodecsAvailable

ErrNoCodecsAvailable indicates that operation isn’t possible because the MediaEngine has no codecs available

ErrUnsupportedCodec

ErrUnsupportedCodec indicates the remote peer doesn’t support the requested codec

ErrSenderWithNoCodecs

ErrSenderWithNoCodecs indicates that a RTPSender was created without any codecs. To send media the MediaEngine needs at least one configured codec.

ErrRTPSenderNewTrackHasIncorrectKind

ErrRTPSenderNewTrackHasIncorrectKind indicates that the new track is of a different kind than the previous/original

ErrRTPSenderNewTrackHasIncorrectEnvelope

ErrRTPSenderNewTrackHasIncorrectEnvelope indicates that the new track has a different envelope than the previous/original

ErrRTPSenderDataSent

ErrRTPSenderDataSent indicates that the sequence number transformer tries to be enabled after the data sending began

ErrRTPSenderSeqTransEnabled

ErrRTPSenderSeqTransEnabled indicates that the sequence number transformer has been already enabled

ErrUnbindFailed

ErrUnbindFailed indicates that a TrackLocal was not able to be unbind

ErrNoPayloaderForCodec

ErrNoPayloaderForCodec indicates that the requested codec does not have a payloader

ErrRegisterHeaderExtensionInvalidDirection

ErrRegisterHeaderExtensionInvalidDirection indicates that a extension was registered with different directions for two different calls.

ErrRegisterHeaderExtensionNoFreeID

ErrRegisterHeaderExtensionNoFreeID indicates that there was no extension ID available which in turn means that all 15 available id(1 through 14) have been used.

ErrSimulcastProbeOverflow

ErrSimulcastProbeOverflow indicates that too many Simulcast probe streams are in flight and the requested SSRC was ignored

ErrDetachNotEnabled

ErrDetachBeforeOpened

ErrDtlsTransportNotStarted

ErrDtlsKeyExtractionFailed

ErrFailedToStartSRTP

ErrFailedToStartSRTCP

ErrInvalidDTLSStart

ErrNoRemoteCertificate

ErrIdentityProviderNotImplemented

ErrNoMatchingCertificateFingerprint

ErrUnsupportedFingerprintAlgorithm

ErrICEConnectionNotStarted

ErrICECandidateTypeUnknown

ErrICEInvalidConvertCandidateType

ErrICEAgentNotExist

ErrICECandidatesConversionFailed

ErrICERoleUnknown

ErrICEProtocolUnknown

ErrICEGathererNotStarted

ErrNetworkTypeUnknown

ErrSDPDoesNotMatchOffer

ErrSDPDoesNotMatchAnswer

ErrPeerConnSDPTypeInvalidValue

ErrPeerConnStateChangeInvalid

ErrPeerConnStateChangeUnhandled

ErrPeerConnSDPTypeInvalidValueSetLocalDescription

ErrPeerConnRemoteDescriptionWithoutMidValue

ErrPeerConnRemoteDescriptionNil

ErrPeerConnSingleMediaSectionHasExplicitSSRC

ErrPeerConnRemoteSSRCAddTransceiver

ErrPeerConnSimulcastMidRTPExtensionRequired

ErrPeerConnSimulcastStreamIDRTPExtensionRequired

ErrPeerConnSimulcastIncomingSSRCFailed

ErrPeerConnStatsCollectionFailed

ErrPeerConnAddTransceiverFromKindOnlyAcceptsOne

ErrPeerConnAddTransceiverFromTrackOnlyAcceptsOne

ErrPeerConnAddTransceiverFromKindSupport

ErrPeerConnAddTransceiverFromTrackSupport

ErrPeerConnSetIdentityProviderNotImplemented

ErrPeerConnWriteRTCPOpenWriteStream

ErrPeerConnTransceiverMidNil

ErrRTPReceiverDTLSTransportNil

ErrRTPReceiverReceiveAlreadyCalled

ErrRTPReceiverWithSSRCTrackStreamNotFound

ErrRTPReceiverForSSRCTrackStreamNotFound

ErrRTPReceiverForRIDTrackStreamNotFound

ErrRTPReceiverStateChangeInvalid

Fields

from: State

to: State

ErrRTPSenderTrackNil

ErrRTPSenderStopped

ErrRTPSenderTrackRemoved

ErrRTPSenderRidNil

ErrRTPSenderNoBaseEncoding

ErrRTPSenderBaseEncodingMismatch

ErrRTPSenderRIDCollision

ErrRTPSenderNoTrackForRID

ErrRTPSenderNil

ErrRTPReceiverNil

ErrRTPSenderDTLSTransportNil

ErrRTPSenderSendAlreadyCalled

ErrRTPTransceiverCannotChangeMid

ErrRTPTransceiverSetSendingInvalidState

ErrRTPTransceiverCodecUnsupported

ErrSCTPTransportDTLS

ErrSDPZeroTransceivers

ErrSDPMediaSectionMediaDataChanInvalid

ErrSDPMediaSectionMultipleTrackInvalid

ErrSettingEngineSetAnsweringDTLSRole

ErrSignalingStateCannotRollback

ErrSignalingStateProposedTransitionInvalid

Fields

from: RTCSignalingState

applying: RTCSdpType

is_local: bool

ErrStatsICECandidateStateInvalid

ErrICETransportNotInNew

ErrCertificatePEMFormatError

ErrSCTPNotEstablished

ErrClosedPipe

ErrInterceptorNotBind

ErrExcessiveRetries

ErrRTPTooShort

Util(Error)

Ice(Error)

Srtp(Error)

Dtls(Error)

Data(Error)

Sctp(Error)

Sdp(Error)

Interceptor(Error)

Rtcp(Error)

Rtp(Error)

Utf8(FromUtf8Error)

RcGen(Error)

MpscSend(String)

ParseInt(ParseIntError)

ParseUrl(ParseError)

InvalidPEM(String)

Error parsing a given PEM string.

new(String)