webrtc

Module ice_transport

The utilities defining transport between peers. Contains ice_transport::ice_server::RTCIceServer struct which describes how peer does ICE (Interactive Connectivity Establishment).

Modules

ice_candidate

ice_candidate_pair

ice_candidate_type

ice_connection_state

ice_gatherer

ice_gatherer_state

ice_gathering_state

ice_parameters

ice_protocol

ice_role

ice_server

ice_transport_state

Structs

RTCIceTransport

ICETransport allows an application access to information about the ICE transport over which packets are sent and received.

Type Aliases

OnConnectionStateChangeHdlrFn

OnSelectedCandidatePairChangeHdlrFn