webrtc::mux

Module mux_func

Functions

match_all

match_all always returns true

match_dtls

MatchFuncs as described in RFC7983 https://tools.ietf.org/html/rfc7983 +––––––––+ | [0..3] -+–> forward to STUN | | | [16..19] -+–> forward to ZRTP | | packet –> | [20..63] -+–> forward to DTLS | | | [64..79] -+–> forward to TURN Channel | | | [128..191] -+–> forward to RTP/RTCP +––––––––+ match_dtls is a MatchFunc that accepts packets with the first byte in [20..63] as defined in RFC7983

match_range

match_range is a MatchFunc that accepts packets with the first byte in [lower..upper]

match_srtcp

match_srtcp is a MatchFunc that only matches SRTCP and not SRTP

match_srtp

match_srtp is a MatchFunc that only matches SRTP and not SRTCP

match_srtp_or_srtcp

Type Aliases

MatchFunc

MatchFunc allows custom logic for mapping packets to an Endpoint