webrtc

Module peer_connection

peer_connection::RTCPeerConnection allows to establish connection between two peers given RTC configuration. Its API is similar to one in JavaScript.

Modules

certificate

configuration

offer_answer_options

Custom media-related options, such as voice_activity_detection, which are negotiated while establishing connection.

peer_connection_state

policy

sdp

RTCSessionDescription - wrapper for SDP text and negotiations stage (RTCSdpType: offer - pranswer - answer - rollback).

signaling_state

Structs

RTCPeerConnection

PeerConnection represents a WebRTC connection that establishes a peer-to-peer communications with another PeerConnection instance in a browser, or to another endpoint implementing the required protocols.

Functions

math_rand_alpha

math_rand_alpha generates a mathematical random alphabet sequence of the requested length.

Type Aliases

OnDataChannelHdlrFn

OnICEConnectionStateChangeHdlrFn

OnNegotiationNeededHdlrFn

OnPeerConnectionStateChangeHdlrFn

OnSignalingStateChangeHdlrFn

OnTrackHdlrFn