webrtc

Module rtp_transceiver

Modules

rtp_codec

rtp_receiver

rtp_sender

rtp_transceiver_direction

Structs

RTCPFeedback

rtcpfeedback signals the connection to use additional RTCP packet types. https://draft.ortc.org/#dom-rtcrtcpfeedback

RTCRtpCapabilities

RTPCapabilities represents the capabilities of a transceiver https://w3c.github.io/webrtc-pc/#rtcrtpcapabilities

RTCRtpCodingParameters

RTPCodingParameters provides information relating to both encoding and decoding. This is a subset of the RFC since Pion WebRTC doesn’t implement encoding/decoding itself http://draft.ortc.org/#dom-rtcrtpcodingparameters

RTCRtpReceiveParameters

RTPReceiveParameters contains the RTP stack settings used by receivers

RTCRtpRtxParameters

RTPRtxParameters dictionary contains information relating to retransmission (RTX) settings. https://draft.ortc.org/#dom-rtcrtprtxparameters

RTCRtpSendParameters

RTPSendParameters contains the RTP stack settings used by receivers

RTCRtpTransceiver

RTPTransceiver represents a combination of an RTPSender and an RTPReceiver that share a common mid.

RTCRtpTransceiverInit

RTPTransceiverInit dictionary is used when calling the WebRTC function addTransceiver() to provide configuration options for the new transceiver.

Constants

TYPE_RTCP_FB_ACK

TYPE_RTCP_FB_ACK ..

TYPE_RTCP_FB_CCM

TYPE_RTCP_FB_CCM ..

TYPE_RTCP_FB_GOOG_REMB

TYPE_RTCP_FB_GOOG_REMB ..

TYPE_RTCP_FB_NACK

TYPE_RTCP_FB_NACK ..

TYPE_RTCP_FB_TRANSPORT_CC

TYPE_RTCP_FBT_RANSPORT_CC ..

Type Aliases

PayloadType

PayloadType identifies the format of the RTP payload and determines its interpretation by the application. Each codec in a RTP Session will have a different PayloadType https://tools.ietf.org/html/rfc3550#section-3

RTCRtpDecodingParameters

RTPDecodingParameters provides information relating to both encoding and decoding. This is a subset of the RFC since Pion WebRTC doesn’t implement decoding itself http://draft.ortc.org/#dom-rtcrtpdecodingparameters

RTCRtpEncodingParameters

RTPEncodingParameters provides information relating to both encoding and decoding. This is a subset of the RFC since Pion WebRTC doesn’t implement encoding itself http://draft.ortc.org/#dom-rtcrtpencodingparameters

SSRC

SSRC represents a synchronization source A synchronization source is a randomly chosen value meant to be globally unique within a particular RTP session. Used to identify a single stream of media. https://tools.ietf.org/html/rfc3550#section-3

TriggerNegotiationNeededFnOption