webrtc::rtp_transceiver

Struct RTCRtpTransceiver

pub struct RTCRtpTransceiver { /* private fields */ }

RTPTransceiver represents a combination of an RTPSender and an RTPReceiver that share a common mid.

Implementations

impl RTCRtpTransceiver

pub async fn new( receiver: Arc<RTCRtpReceiver>, sender: Arc<RTCRtpSender>, direction: RTCRtpTransceiverDirection, kind: RTPCodecType, codecs: Vec<RTCRtpCodecParameters>, media_engine: Arc<MediaEngine>, trigger_negotiation_needed: TriggerNegotiationNeededFnOption, ) -> Arc<Self>

pub async fn set_codec_preferences( &self, codecs: Vec<RTCRtpCodecParameters>, ) -> Result<()>

set_codec_preferences sets preferred list of supported codecs if codecs is empty or nil we reset to default from MediaEngine

pub async fn sender(&self) -> Arc<RTCRtpSender>

sender returns the RTPTransceiver’s RTPSender if it has one

pub async fn set_sender_track( self: &Arc<Self>, sender: Arc<RTCRtpSender>, track: Option<Arc<dyn TrackLocal + Send + Sync>>, ) -> Result<()>

set_sender_track sets the RTPSender and Track to current transceiver

pub async fn set_sender(self: &Arc<Self>, s: Arc<RTCRtpSender>)

pub async fn receiver(&self) -> Arc<RTCRtpReceiver>

receiver returns the RTPTransceiver’s RTPReceiver if it has one

pub fn mid(&self) -> Option<SmolStr>

mid gets the Transceiver’s mid value. When not already set, this value will be set in CreateOffer or create_answer.

pub fn kind(&self) -> RTPCodecType

kind returns RTPTransceiver’s kind.

pub fn direction(&self) -> RTCRtpTransceiverDirection

direction returns the RTPTransceiver’s desired direction.

pub async fn set_direction(&self, d: RTCRtpTransceiverDirection)

Set the direction of this transceiver. This might trigger a renegotiation.

pub fn current_direction(&self) -> RTCRtpTransceiverDirection

current_direction returns the RTPTransceiver’s current direction as negotiated.

If this transceiver has never been negotiated or if it’s stopped this returns RTCRtpTransceiverDirection::Unspecified.

pub async fn stop(&self) -> Result<()>

stop irreversibly stops the RTPTransceiver