webrtc::sctp_transport

Struct RTCSctpTransport

pub struct RTCSctpTransport { /* private fields */ }

SCTPTransport provides details about the SCTP transport.

Implementations

impl RTCSctpTransport

pub fn transport(&self) -> Arc<RTCDtlsTransport>

transport returns the DTLSTransport instance the SCTPTransport is sending over.

pub fn get_capabilities(&self) -> SCTPTransportCapabilities

get_capabilities returns the SCTPCapabilities of the SCTPTransport.

pub async fn start(&self, _remote_caps: SCTPTransportCapabilities) -> Result<()>

Start the SCTPTransport. Since both local and remote parties must mutually create an SCTPTransport, SCTP SO (Simultaneous Open) is used to establish a connection over SCTP.

pub async fn stop(&self) -> Result<()>

Stop stops the SCTPTransport

pub fn on_error(&self, f: OnErrorHdlrFn)

on_error sets an event handler which is invoked when the SCTP connection error occurs.

pub fn on_data_channel(&self, f: OnDataChannelHdlrFn)

on_data_channel sets an event handler which is invoked when a data channel message arrives from a remote peer.

pub fn on_data_channel_opened(&self, f: OnDataChannelOpenedHdlrFn)

on_data_channel_opened sets an event handler which is invoked when a data channel is opened

pub fn max_channels(&self) -> u16

max_channels is the maximum number of RTCDataChannels that can be open simultaneously.

pub fn state(&self) -> RTCSctpTransportState

state returns the current state of the SCTPTransport