webrtc::track::track_local::track_local_static_rtp

Struct TrackLocalStaticRTP

pub struct TrackLocalStaticRTP { /* private fields */ }

TrackLocalStaticRTP is a TrackLocal that has a pre-set codec and accepts RTP Packets. If you wish to send a media.Sample use TrackLocalStaticSample

Implementations

impl TrackLocalStaticRTP

pub fn new(codec: RTCRtpCodecCapability, id: String, stream_id: String) -> Self

returns a TrackLocalStaticRTP without rid.

pub fn new_with_rid( codec: RTCRtpCodecCapability, id: String, rid: String, stream_id: String, ) -> Self

returns a TrackLocalStaticRTP with rid.

pub fn codec(&self) -> RTCRtpCodecCapability

codec gets the Codec of the track

pub async fn any_binding_paused(&self) -> bool

pub async fn all_binding_paused(&self) -> bool

pub async fn write_rtp_with_extensions( &self, p: &Packet, extensions: &[HeaderExtension], ) -> Result<usize>

write_rtp_with_extensions writes a RTP Packet to the TrackLocalStaticRTP If one PeerConnection fails the packets will still be sent to all PeerConnections. The error message will contain the ID of the failed PeerConnections so you can remove them

If the RTCRtpSender direction is such that no packets should be sent, any call to this function are blocked internally. Care must be taken to not increase the sequence number while the sender is paused. While the actual sending is blocked, the receiver will miss out when the sequence number “rolls over”, which in turn will break SRTP.

Extensions that are already configured on the packet are overwritten by extensions in extensions.

pub async fn write_rtp_with_extensions_attributes( &self, p: &Packet, extensions: &[HeaderExtension], attr: &Attributes, ) -> Result<usize>