webrtc::track::track_remote

Struct TrackRemote

pub struct TrackRemote { /* private fields */ }

TrackRemote represents a single inbound source of media

Implementations

impl TrackRemote

pub fn tid(&self) -> usize

pub fn id(&self) -> String

id is the unique identifier for this Track. This should be unique for the stream, but doesn’t have to globally unique. A common example would be ‘audio’ or ‘video’ and StreamID would be ‘desktop’ or ‘webcam’

pub fn set_id(&self, s: String)

pub fn stream_id(&self) -> String

stream_id is the group this track belongs too. This must be unique

pub fn set_stream_id(&self, s: String)

pub fn rid(&self) -> &str

rid gets the RTP Stream ID of this Track With Simulcast you will have multiple tracks with the same ID, but different RID values. In many cases a TrackRemote will not have an RID, so it is important to assert it is non-zero

pub fn payload_type(&self) -> PayloadType

payload_type gets the PayloadType of the track

pub fn set_payload_type(&self, payload_type: PayloadType)

pub fn kind(&self) -> RTPCodecType

kind gets the Kind of the track

pub fn set_kind(&self, kind: RTPCodecType)

pub fn ssrc(&self) -> SSRC

ssrc gets the SSRC of the track

pub fn set_ssrc(&self, ssrc: SSRC)

pub fn msid(&self) -> String

msid gets the Msid of the track

pub fn codec(&self) -> RTCRtpCodecParameters

codec gets the Codec of the track

pub fn set_codec(&self, codec: RTCRtpCodecParameters)

pub fn params(&self) -> RTCRtpParameters

pub fn set_params(&self, params: RTCRtpParameters)

pub fn onmute<F>(&self, handler: F)
where F: FnMut() -> Pin<Box<dyn Future<Output = ()> + Send + 'static>> + Send + 'static + Sync,

pub fn onunmute<F>(&self, handler: F)
where F: FnMut() -> Pin<Box<dyn Future<Output = ()> + Send + 'static>> + Send + 'static + Sync,

pub async fn read(&self, b: &mut [u8]) -> Result<(Packet, Attributes)>

Reads data from the track.

Cancel Safety: This method is not cancel safe. Dropping the resulting Future before it returns std::task::Poll::Ready will cause data loss.

pub async fn read_rtp(&self) -> Result<(Packet, Attributes)>

read_rtp is a convenience method that wraps Read and unmarshals for you.